+++ packages/api-client/src/index.ts
@@ -1,4 +1,6 @@
 export { ApiClient } from './components/ApiClient'
+export * from './components/ApiClient/Request/RequestAuthentication'
+
 export { default as HttpMethod } from './components/HttpMethod.vue'
 
 export * from './helpers'