+++ packages/api-client/src/types.ts
@@ -1,42 +1,4 @@
-import { type BaseParameter } from '@scalar/oas-utils'
-
-export type BasicAuth = {
-  username: string
-  password: string
-  active: boolean
-}
-
-export type OAuthTwo = {
-  generatedToken: string
-  discoveryURL: string
-  authURL: string
-  accessTokenURL: string
-  clientID: string
-  clientSecret: string
-  scope: string
-  active: boolean
-}
-
-export type Bearer = {
-  token: string
-  active: boolean
-}
-
-export type Digest = {
-  username: string
-  password: string
-  active: boolean
-}
-
-export type AuthType = 'basic' | 'oauthTwo' | 'bearer' | 'digest' | 'none'
-
-export type AuthState = {
-  type: AuthType
-  basic: BasicAuth
-  oauthTwo: OAuthTwo
-  bearer: Bearer
-  digest: Digest
-}
+import type { BaseParameter } from '@scalar/oas-utils'
 
 export type Header = BaseParameter
 
@@ -45,6 +7,11 @@ export type Query = BaseParameter
 export type Cookie = BaseParameter
 
 export type FormDataItem = BaseParameter
+
+export type GeneratedParameter = {
+  name: string
+  value: string
+}
 
 /** Complete request state for a client request */
 export type ClientRequestConfig = {