+++ packages/api-reference-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @scalar/api-reference-react
 
+## 0.1.19
+
+### Patch Changes
+
+- Updated dependencies [d70bda7]
+  - @scalar/api-reference@1.20.11
+
 ## 0.1.18
 
 ### Patch Changes